+++ docs/pages/quickstart/first_experiments/running_a_real_policy/openpi.rst
@@ -86,10 +86,6 @@ To measure success rates across several variations of the environment in a singl
      --viz kit \
      --experiment_config isaaclab_arena_environments/experiment_configs/droid_pnp_srl_openpi_experiment.yaml
 
-The Experiment config adds cameras to each environment. The runner reads that requirement from the
-config before starting Isaac Sim, so camera support is enabled for you and ``--enable_cameras``
-is not needed.
-
 This runs nine jobs sequentially — each varying the object, background, and destination — and reports a per-job success rate.
 Each evaluation is run without restarting Isaac Sim to save on the startup time.
